Move the `Version` class into `codegen-core`
