+++ codegen/src/test/kotlin/software/amazon/smithy/rust/codegen/smithy/generators/CargoTomlGeneratorTest.kt
@@ -6,9 +6,9 @@
 package software.amazon.smithy.rust.codegen.smithy.generators
 
 import org.junit.jupiter.api.Test
+import software.amazon.smithy.rust.codegen.core.Version
 import software.amazon.smithy.rust.codegen.rustlang.CargoDependency
 import software.amazon.smithy.rust.codegen.rustlang.CratesIo
-import software.amazon.smithy.rust.codegen.smithy.Version
 import software.amazon.smithy.rust.codegen.testutil.TestWorkspace
 import software.amazon.smithy.rust.codegen.testutil.compileAndTest
 import software.amazon.smithy.rust.codegen.testutil.unitTest
